[chore] statistics menu file move study folder to statistics folder
