+++ src/main/frontend/src/index.js
@@ -7,10 +7,10 @@ import NotFound from "./NotFound";
 import AbsentSchedule from "./pages/absent-schedule/AbsentSchedule";
 import MyPage from "./pages/my-page/MyPage";
 import NoticeBoard from "./pages/notice-board/NoticeBoard";
-import AttendanceStatistics from "./pages/study/AttendanceStatistics";
+import AttendanceStatistics from "./pages/statistics/AttendanceStatistics";
 import Management from "./pages/study/Management";
 import MembersInfo from "./pages/home/MembersInfo";
-import WakeUpSuccessRate from "./pages/study/WakeUpSuccessRate";
+import WakeUpSuccessRate from "./pages/statistics/WakeUpSuccessRate";
 import Login from "./pages/login/login";
 import Join from "./pages/join/join";
 import CssBaseline from "@mui/material/CssBaseline";
