+++ src/main/frontend/src/pages/statistics/WakeUpSuccessRate.js
